requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
openai>=1.51.0
chromadb>=0.5.4
sentence-transformers>=3.2.0
FlagEmbedding>=1.2.10
PyYAML>=6.0
